Fix type conversion warnings.

+++ src/drw_entities.h
@@ -918,7 +918,7 @@ public:
     }
 
     void update() {
-        numedges = objlist.size();
+        numedges = (int)objlist.size();
     }
 
 public:
